+++ src/api/axios-instance.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-// todo: I don't like this cyclic dependency, but struggling to find a better way atm.
+// I don't like this cyclic dependency, but struggling to find a better way atm.
 import store from '@/store/index';
 
 // eslint-disable-next-line import/no-mutable-exports
@@ -16,13 +16,18 @@ export function setInstance(url, token) {
   axiosInstance.interceptors.response.use(
     (response) => response,
     (err) => {
-      if (!err.response) {
+      if (!err?.response) {
         store.dispatch('application/connectionError');
       }
-      if (err.response.status === 401) {
+
+      if (err?.response?.status === 401 || err?.response?.status === 403) {
         store.dispatch('application/logout');
       }
       return Promise.reject(err);
     },
   );
+}
+
+export function handleError(error) {
+  return error?.response?.data?.detail || error;
 }