+++ src/api/stager-api.js
@@ -1,4 +1,4 @@
-import { axiosInstance as axios } from '@/api/axios-instance';
+import { axiosInstance as axios, handleError } from '@/api/axios-instance';
 
 /**
  * Returns a single stager.
@@ -6,7 +6,7 @@ import { axiosInstance as axios } from '@/api/axios-instance';
 export function getStager(id) {
   return axios.get(`/stagers/${id}`)
     .then(({ data }) => data)
-    .catch((error) => Promise.reject(error.response.data.detail));
+    .catch((error) => Promise.reject(handleError(error)));
 }
 
 /**
@@ -15,7 +15,7 @@ export function getStager(id) {
 export function getStagers() {
   return axios.get('/stagers')
     .then(({ data }) => data.records)
-    .catch((error) => Promise.reject(error.response.data.detail));
+    .catch((error) => Promise.reject(handleError(error)));
 }
 
 /**
@@ -25,7 +25,7 @@ export function getStagers() {
 export function getStagerTemplate(templateId) {
   return axios.get(`/stager-templates/${templateId}`)
     .then(({ data }) => data)
-    .catch((error) => Promise.reject(error.response.data.detail));
+    .catch((error) => Promise.reject(handleError(error)));
 }
 
 /**
@@ -35,7 +35,7 @@ export function getStagerTemplate(templateId) {
 export function createStager(template, name, options) {
   return axios.post('/stagers', { name, template, options })
     .then(({ data }) => data)
-    .catch((error) => Promise.reject(error.response.data.detail));
+    .catch((error) => Promise.reject(handleError(error)));
 }
 
 /**
@@ -45,7 +45,7 @@ export function createStager(template, name, options) {
 export function updateStager(id, options) {
   return axios.put(`/stagers/${id}`, options)
     .then(({ data }) => data)
-    .catch((error) => Promise.reject(error.response ? error.response.data.detail : error));
+    .catch((error) => Promise.reject(handleError(error)));
 }
 
 /**
@@ -54,7 +54,7 @@ export function updateStager(id, options) {
 export function getStagerTemplates() {
   return axios.get('/stager-templates')
     .then(({ data }) => data.records)
-    .catch((error) => Promise.reject(error.response.data.detail));
+    .catch((error) => Promise.reject(handleError(error)));
 }
 
 /**
@@ -64,7 +64,7 @@ export function getStagerTemplates() {
 export function getStagerByName(name) {
   return axios.get(`/stagers/${name}`)
     .then(({ data }) => data.stagers[0])
-    .catch((error) => Promise.reject(error.response.data.error));
+    .catch((error) => Promise.reject(handleError(error)));
 }
 
 /**
@@ -73,5 +73,5 @@ export function getStagerByName(name) {
  */
 export function deleteStager(id) {
   return axios.delete(`/stagers/${id}`)
-    .catch((error) => Promise.reject(error.response.data.detail));
+    .catch((error) => Promise.reject(handleError(error)));
 }